refactor names in frontend part
