+++ ui/src/components/category-item/category-item.scss
