+++ ui/src/components/directory/directory.jsx
@@ -1,6 +1,6 @@
-import CategoryItem from '../category-item/category-item.component';
+import CategoryItem from '../category-item/category-item';
 
-import './directory.styles.scss';
+import './directory.scss';
 
 const Directory = ({ categories }) => {
   return (
