+++ ui/src/components/movie-by-id/movie-by-id.css
