+++ ui/src/components/movie-list/movie-list.css
@@ -1,0 +1,5 @@
+.movie-list {
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+}
