--- ui/src/components/movie-list/movie-list.styles.css
@@ -1,5 +1,0 @@
-.movie-list {
-    background-color: white;
-    display: flex;
-    flex-direction: column;
-  }
