+++ ui/src/components/movie-little-card/movie-little-card.css
