+++ ui/src/components/sidebar/sidebar.css
