+++ ui/src/routes/navigation/navigation.css
